ms-reserva/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
drf-spectacular==0.28.0
django-environ==0.12.0
mysqlclient==2.2.7